Add ability to set output format
Before this commit the output message payload was always
just the plain text value of the received reply.  This commit
adds the ability to set the output format, either to plain/text
(which is the default) or application/json.  In the latter case
the output message payload is set as { name: { value: value }},
which is directly compatible with the Ell-i Fiware node.

That allows a coap-request node to be directly routed to a
Fiware device-out node, thereby making a trivial CoAP-to-Fiware
IoT agent.